Allow vms connected with a router access to eips on this router
According to the current nat rules, vms connected with a router
can not access to the eips on this router, because rules in nat
won't process packets from guest network.

This patch changes the rules to make it accessible.

Besides, found a bug about snat rule of eip, also fixed in this patch.

For zxwing/premium#1638